telegramBot/requirements.txt
requests
python-dotenv
paho-mqtt
flask
flask-cors
schedule
